Fix dangling pointers, narrowing casts, and duplicated code
- Replace raw entity pointers in debug_stats with entity IDs to prevent
  use-after-free on entity removal/map transitions
- Widen experience fields from uint32_t to int64_t to match server protocol
  and prevent silent truncation above 4.29B
- Extract equip_slot_from_server() to deduplicate slot mapping switch
- Extract shared spawn helpers (init_entity_transform, init_entity_visual_type,
  init_entity_dead_state) from 3 entity spawn handlers
- Add missing contains("data") guards in handle_view_range_update and
  handle_command_response
- Remove debug logging that dumped full skills_data JSON at info level

+++ src/entity/components.hpp
@@ -239,9 +239,9 @@ struct stats_component
 
     // Level/XP
     uint16_t level = 1;
-    uint32_t experience = 0;
-    uint32_t exp = 0; // Alias for experience
-    uint32_t experience_to_next = 100;
+    int64_t experience = 0;
+    int64_t exp = 0; // Alias for experience
+    int64_t experience_to_next = 100;
 
     // Combat stats
     int32_t attack_power = 10;
